requirements-file.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
python-dotenv==1.0.1
openai==1.14.0
pytest==7.4.3
httpx==0.26.0
python-multipart==0.0.9
pymongo==4.6.1
